+++ client/src/styles/features/advancedSearch/AdvancedSearchContainers.ts
@@ -18,7 +18,6 @@ export const StyledContainer = styled(Col)<IProps>`
 
   &.advancedSearchBody {
     padding-left: 1rem;
-    box-shadow: 0px 4px 5px 0px ${theme.color.black20};
     background: ${theme.color.white};
     border-top-left-radius: ${(props) =>
       props.$asBodyBorderTopLeftRadius || theme.border.radius};
@@ -31,7 +30,6 @@ export const StyledContainer = styled(Col)<IProps>`
   &.helpText {
     padding-right: 1rem;
     padding-left: 1rem;
-    box-shadow: 4px 4px 5px 0px ${theme.color.black20};
     background: ${theme.color.white};
     border-top-right-radius: ${(props) =>
       props.$helpTextBorderTopRightRadius || theme.border.radius};